+++ src/main/java/owl/tree/rmfarma/manufacture/domain/data/masterorderdetails/OrderDetailCreateResourceDto.java
@@ -15,7 +15,7 @@ public class OrderDetailCreateResourceDto {
     private String productName;
     private String productCode;
     private String productLaboratory;
-    private Integer quantity;
+    private Double quantity;
     private String complementName;
     private String complementCode;
     private String product;
@@ -25,7 +25,7 @@ public class OrderDetailCreateResourceDto {
     private String viaDescription;
     private String via;
     private String prot;
-    private Integer volumeTotal;
+    private Double volumeTotal;
     private String unitMetric;
     private String condition;
     private String administrationTime;
